src/requirements.txt
Django>=2.2.0
django-rest-framework==0.1.0
psycopg2>=2.8.3
django-filter==2.1.0
django-cors-headers==3.0.2
flake8>=3.7.8
coverage>=4.5.3